refractored consumer api
